+++ conda/base/constants.py
@@ -94,14 +94,11 @@ class LinkType(Enum):
         return self.name.replace('_', '-')
 
 
-
-
 PREFIX_PLACEHOLDER = ('/opt/anaconda1anaconda2'
                       # this is intentionally split into parts,
                       # such that running this program on itself
                       # will leave it unchanged
                       'anaconda3')
-
 
 machine_bits = 8 * tuple.__itemsize__
